# postman/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Intern Kominfo API
  version: 1.1.0
  description: |
    API for the Intern Kominfo application — a government intern management
    platform with Work-From-Home (WFH) reporting, change management workflows,
    user administration, and QR-based document verification.

    Authentication uses Laravel Sanctum bearer tokens.
    Authorization uses Spatie permission-based roles (admin, kepala_tim, staf).
servers:
  - url: http://localhost
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Email atau password salah.
...
    PaginationMeta:
      type: object
      required: [current_page, last_page, total]
      properties:
        current_page:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 5
        total:
          type: integer
          example: 72
    # ── Auth ────────────────────────────────────────────────────────
...
    ReportUserNested:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Budi Santoso
        nip:
          type: string
          example: "199001012023011001"
        position:
          type: string
          nullable: true
          example: Staf Pelaksana
        rank:
          type: string
          nullable: true
          example: Pengatur Muda
        signature_path:
          type: string
          nullable: true
          example: signatures/1-signature.png
      example:
        id: 1
        name: Budi Santoso
        nip: "199001012023011001"
        position: Staf Pelaksana
        rank: Pengatur Muda
        signature_path: signatures/1-signature.png
...
    StoreInitiationRequest:
      type: object
      required: [field_id, description, reason]
      properties:
        field_id:
          type: integer
          example: 1
        needed_by_date:
          type: string
          format: date
          nullable: true
          example: "2025-08-15"
        description:
          type: string
          example: "Migrasi server aplikasi dari on-premise ke cloud kominfo"
        reason:
          type: string
          example: "Server on-premise saat ini sudah tidak memadai untuk menangani lonjakan trafik pengguna dan sering mengalami downtime."
      example:
        field_id: 1
        needed_by_date: "2025-08-15"
        description: "Migrasi server aplikasi dari on-premise ke cloud kominfo"
        reason: "Server on-premise saat ini sudah tidak memadai untuk menangani lonjakan trafik pengguna dan sering mengalami downtime."
    InitiationResource:
      type: object
      required: [id, field_id, initiator_id, doc_number, status]
      properties:
        id:
          type: integer
          example: 8
        field_id:
          type: integer
          example: 1
        initiator_id:
          type: integer
          example: 1
        reviewer_id:
          type: integer
          nullable: true
          example: 5
        doc_number:
          type: string
          example: CHG-2025-008
        initiation_date:
          type: string
          format: date
          nullable: true
          example: "2025-07-14"
        needed_by_date:
          type: string
          format: date
          nullable: true
          example: "2025-08-15"
        description:
          type: string
          example: "Migrasi server aplikasi dari on-premise ke cloud kominfo"
        reason:
          type: string
          example: "Server on-premise saat ini sudah tidak memadai untuk menangani lonjakan trafik pengguna dan sering mengalami downtime."
        status:
          type: string
          enum: [draft, pending, approved, rejected]
          example: pending
        review_status:
          type: string
          nullable: true
          example: null
        reviewed_at:
          type: string
          format: date-time
          nullable: true
          example: null
        review_reason:
          type: string
          nullable: true
          example: null
        initiator_signed_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-07-14T10:30:00.000000Z"
        created_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-07-14T10:00:00.000000Z"
        field:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: Bidang Teknologi Informasi
        initiator:
          $ref: "#/components/schemas/ReportUserNested"
        implementations:
          type: array
          items:
            $ref: "#/components/schemas/ImplementationResource"
      example:
        id: 8
        field_id: 1
        initiator_id: 1
        reviewer_id: 5
        doc_number: CHG-2025-008
        initiation_date: "2025-07-14"
        needed_by_date: "2025-08-15"
        description: "Migrasi server aplikasi dari on-premise ke cloud kominfo"
        reason: "Server on-premise saat ini sudah tidak memadai untuk menangani lonjakan trafik pengguna dan sering mengalami downtime."
        status: pending
        review_status: null
        reviewed_at: null
        review_reason: null
        initiator_signed_at: "2025-07-14T10:30:00.000000Z"
        created_at: "2025-07-14T10:00:00.000000Z"
        field:
          id: 1
          name: Bidang Teknologi Informasi
        initiator:
          id: 1
          name: Budi Santoso
          nip: "199001012023011001"
          position: Staf Pelaksana
          rank: Pengatur Muda
          signature_path: signatures/1-signature.png
        implementations: []
    InitiationListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/InitiationResource"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        success: true
        data:
          - id: 8
            field_id: 1
            initiator_id: 1
            reviewer_id: 5
            doc_number: CHG-2025-008
            initiation_date: "2025-07-14"
            needed_by_date: "2025-08-15"
            description: "Migrasi server aplikasi dari on-premise ke cloud kominfo"
            reason: "Server on-premise saat ini sudah tidak memadai untuk menangani lonjakan trafik pengguna dan sering mengalami downtime."
            status: pending
            review_status: null
            reviewed_at: null
            review_reason: null
            initiator_signed_at: "2025-07-14T10:30:00.000000Z"
            created_at: "2025-07-14T10:00:00.000000Z"
            field:
              id: 1
              name: Bidang Teknologi Informasi
            initiator:
              id: 1
              name: Budi Santoso
              nip: "199001012023011001"
              position: Staf Pelaksana
              rank: Pengatur Muda
              signature_path: signatures/1-signature.png
            implementations: []
        meta:
          current_page: 1
          last_page: 1
          total: 3
...
    ImplementationResource:
      type: object
      required: [id, change_initiation_id, status]
      properties:
        id:
          type: integer
          example: 4
        change_initiation_id:
          type: integer
          example: 8
        priority:
          type: string
          nullable: true
          example: high
        impact:
          type: string
          nullable: true
          example: high
        production_impact:
          type: string
          nullable: true
          example: "Aplikasi akan tidak dapat diakses selama 2 jam pada saat migrasi data."
        required_effort:
          type: string
          nullable: true
          example: "3 orang x 5 hari kerja"
        cost_needed:
          type: boolean
          nullable: true
          example: true
        cost_amount:
          type: number
          format: double
          nullable: true
          example: 15000000
        resources:
          type: string
          nullable: true
          example: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps"
        test_plan:
          type: string
          nullable: true
          example: "Smoke test semua endpoint API, load test dengan 100 concurrent users, uji rollback ke server lama."
        evaluator_id:
          type: integer
          nullable: true
          example: 5
        evaluator_signed_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-07-14T11:00:00.000000Z"
        review_status:
          type: string
          nullable: true
          example: null
        review_response:
          type: string
          nullable: true
          example: null
        execution_date:
          type: string
          format: date
          nullable: true
          example: null
        reviewer_id:
          type: integer
          nullable: true
          example: null
        reviewer_signed_at:
          type: string
          format: date-time
          nullable: true
          example: null
        implementation_result:
          type: string
          nullable: true
          example: null
        testing_result:
          type: string
          nullable: true
          example: null
        release_date:
          type: string
          format: date
          nullable: true
          example: null
        responsible_id:
          type: integer
          nullable: true
          example: null
        responsible_signed_at:
          type: string
          format: date-time
          nullable: true
          example: null
        status:
          type: string
          enum: [draft, submitted, completed, rejected, revision]
          example: submitted
        change_types:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 1
              name:
                type: string
                example: Perubahan Infrastruktur
        attachments:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 15
              path:
                type: string
                example: implementations/4/diagram-arsitektur.png
              url:
                type: string
                format: uri
                example: http://localhost/storage/implementations/4/diagram-arsitektur.png
              sort_order:
                type: integer
                example: 1
      example:
        id: 4
        change_initiation_id: 8
        priority: high
        impact: high
        production_impact: "Aplikasi akan tidak dapat diakses selama 2 jam pada saat migrasi data."
        required_effort: "3 orang x 5 hari kerja"
        cost_needed: true
        cost_amount: 15000000
        resources: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps"
        test_plan: "Smoke test semua endpoint API, load test dengan 100 concurrent users, uji rollback ke server lama."
        evaluator_id: 5
        evaluator_signed_at: "2025-07-14T11:00:00.000000Z"
        review_status: null
        review_response: null
        execution_date: null
        reviewer_id: null
        reviewer_signed_at: null
        implementation_result: null
        testing_result: null
        release_date: null
        responsible_id: null
        responsible_signed_at: null
        status: submitted
        change_types:
          - id: 1
            name: Perubahan Infrastruktur
          - id: 3
            name: Perubahan Data
        attachments:
          - id: 15
            path: implementations/4/diagram-arsitektur.png
            url: http://localhost/storage/implementations/4/diagram-arsitektur.png
            sort_order: 1
...
security:
  - bearerAuth: []
...
paths:
...
  /api/changes/initiations:
    get:
      tags: [Change Initiation]
      summary: List change initiations (paginated)
      operationId: listInitiations
      parameters:
        - name: per_page
          in: query
          schema:
            type: integer
            default: 15
            maximum: 100
          example: 15
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, approved, rejected]
          example: pending
        - name: field_id
          in: query
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Paginated initiation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiationListResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Unauthenticated.
        "403":
          description: Missing permission change.initiation.view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: This action is unauthorized.
    post:
      tags: [Change Initiation]
      summary: Create a change initiation
      operationId: createInitiation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInitiationRequest"
            example:
              field_id: 1
              needed_by_date: "2025-08-15"
              description: "Migrasi server aplikasi dari on-premise ke cloud kominfo"
              reason: "Server on-premise saat ini sudah tidak memadai untuk menangani lonjakan trafik pengguna dan sering mengalami downtime."
      responses:
        "201":
          description: Initiation created
          content:
            application/json:
              schema:
                type: object
                required: [success, message, data]
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Inisiasi perubahan berhasil dibuat.
                  data:
                    $ref: "#/components/schemas/InitiationResource"
              example:
                success: true
                message: Inisiasi perubahan berhasil dibuat.
                data:
                  id: 8
                  field_id: 1
                  initiator_id: 1
                  reviewer_id: null
                  doc_number: CHG-2025-008
                  initiation_date: "2025-07-14"
                  needed_by_date: "2025-08-15"
                  description: "Migrasi server aplikasi dari on-premise ke cloud kominfo"
                  reason: "Server on-premise saat ini sudah tidak memadai untuk menangani lonjakan trafik pengguna dan sering mengalami downtime."
                  status: draft
                  review_status: null
                  reviewed_at: null
                  review_reason: null
                  initiator_signed_at: null
                  created_at: "2025-07-14T10:00:00.000000Z"
                  field:
                    id: 1
                    name: Bidang Teknologi Informasi
                  initiator:
                    id: 1
                    name: Budi Santoso
                    nip: "199001012023011001"
                    position: Staf Pelaksana
                    rank: Pengatur Muda
                    signature_path: signatures/1-signature.png
                  implementations: []
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Unauthenticated.
        "403":
          description: Missing permission change.initiation.create
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: This action is unauthorized.
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: "The description field is required."
...